refactor: seal library internals behind pub(crate)
Demote parser, render, and defaults modules from pub to pub(crate) so
internal refactors are no longer semver-visible. The crate's own fixture
integration tests keep access through a #[doc(hidden)] internal facade
re-exporting Parser, DocxParser, PptxParser, XlsxParser, TypstOutput,
and generate_typst.

Sealing exposed dead code that pub visibility previously masked:
- remove the native detect_missing_font_fallbacks wrapper (only the wasm
  pipeline branch ever called the path-based variant)
- allow dead_code on native-only font discovery items for wasm builds
  (embedded_fonts, font_context, font_subst, FontSource::Shared)
- gate DEFAULT_STREAMING_CHUNK_SIZE dead_code allow on the pdf-ops
  feature that its only consumer requires
- cfg-gate Path/Mutex imports used only by native code paths

Verified warning-free: workspace clippy -D warnings, wasm32 check (with
and without the wasm feature), and --no-default-features check.
