+++ crates/office2pdf/tests/docx_fixtures.rs
@@ -10,13 +10,13 @@ mod common;
 use std::path::PathBuf;
 
 use office2pdf::config::ConvertOptions;
+use office2pdf::internal::DocxParser;
+use office2pdf::internal::Parser;
 use office2pdf::ir::{
     ArrowHead, Block, BorderLineStyle, Color, FlowPage, FrameAnchor, HFInline, Insets, ListKind,
     Page, Paragraph, PositionedTabAlignment, PositionedTabRelativeTo, Run, ShapeKind,
     TextBoxVerticalAlign,
 };
-use office2pdf::parser::Parser;
-use office2pdf::parser::docx::DocxParser;
 
 // ---------------------------------------------------------------------------
 // Helpers
