+++ crates/office2pdf/tests/pptx_fixtures.rs
@@ -10,10 +10,10 @@ mod common;
 use std::path::PathBuf;
 
 use office2pdf::config::ConvertOptions;
+use office2pdf::internal::Parser;
+use office2pdf::internal::PptxParser;
+use office2pdf::internal::generate_typst;
 use office2pdf::ir::{Block, Color, FixedElementKind, FixedPage, Page};
-use office2pdf::parser::Parser;
-use office2pdf::parser::pptx::PptxParser;
-use office2pdf::render::typst_gen::generate_typst;
 
 // ---------------------------------------------------------------------------
 // Helpers
